#41 Add comment explaining need for artificial isSaving time
