+++ frontend/src/providers/SavingContextProvider.tsx
@@ -1,6 +1,6 @@
 import { useDispatch } from 'react-redux';
+import { ReactNode, useState } from 'react';
 import { AppDispatch } from '../redux/store';
-import { ReactNode, useEffect, useState } from 'react';
 import { SavingContext } from '../contexts/SavingContext';
 import { saveSelectedMod as reduxSaveSelectedMod } from '../redux/slices/saving';
 
@@ -8,18 +8,9 @@ export default function SavingContextProvider({ children }: { children: ReactNod
     const dispatch = useDispatch<AppDispatch>();
 
     const [lastSaved, setLastSaved] = useState(Date.now());
-    const [isSaving, setIsSaving] = useState(false);
     const [error, setError] = useState<string | null>(null);
 
-    useEffect(() => {
-        if (isSaving) {
-            const timeout = setTimeout(() => setIsSaving(false), 350);
-            return () => clearTimeout(timeout);
-        }
-    }, [isSaving]);
-
     const save = () => {
-        setIsSaving(true);
         setError(null);
         setLastSaved(Date.now());
     };
@@ -41,7 +32,6 @@ export default function SavingContextProvider({ children }: { children: ReactNod
         <SavingContext.Provider
             value={{
                 lastSaved,
-                isSaving,
                 error,
                 save,
                 saveSelectedMod,
